LPS-67460 Rename module
