--- modules/sdk/ant-gradle-exec/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Ant Gradle Exec
-Bundle-SymbolicName: com.liferay.ant.gradle.exec
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.ant.gradle.exec
