+++ modules/sdk/ant-gradle/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Ant Gradle
+Bundle-SymbolicName: com.liferay.ant.gradle
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.ant.gradle
