+++ modules/sdk/source-formatter.properties
@@ -8,7 +8,7 @@
     diamond.operator.excludes=\
         **/ant-beanshell/**,\
         **/ant-bnd/**,\
-        **/ant-gradle-exec/**
+        **/ant-gradle/**
 
     source.formatter.excludes=\
         modules/sdk/ant-bnd/src/test/resources/com/liferay/ant/bnd/jsp/dependencies/*.jsp,\
